Made option to remove "temporary" information like the time of the day and the amount of free memory.
This makes it possible to diff two successive runs of system-info to see how a system has changed.
For system-info.d scripts, a new variable "temporaries" is set to 1 (the default) to show temporary
information.  If set to 0, temporary information should be avoided in the output.

Temporary information is defined as information which changes as a side effect of normal system
operation, like the fill size of buffers, available memory or disk, the number or state of network
connections and so on.