Two new flabbergasting features!
1) system-info will now list all current outgoing connections from the
current host along with which process it is (with process name) and
how many connections that process has open to the same target (if more
than one, like DB connections).

  List of cochise's outgoing connections:
     - 7162/opera 2 conns -> 233.12.120.177:8080
     - 7162/opera 2 conns -> 233.12.120.177:8080
     - 9455/ssh 1 conns -> 551.11.13.2:22
     - 7162/opera 1 conns -> 233.12.120.177:8080
     - 6058/dropbox 1 conns -> 199.47.218.147:80
     - 32241/ssh 1 conns -> 192.168.1.233:2343
     - 2730/ssh 1 conns -> 192.168.1.233:2343
     - 2500/firefox-bin 1 conns -> 76.27.219.195:443
     - 2500/firefox-bin 1 conns -> 89.19.192.41:443
     - 2500/firefox-bin 1 conns -> 873.194.32.53:443
     - 23495/ssh 1 conns -> 192.168.1.233:2343
     - 23084/ssh 1 conns -> 192.168.1.233:2343
     - 21847/ssh 1 conns -> 192.168.1.233:2343
     - 20914/ssh 1 conns -> 192.168.1.233:2343
     - 15926/ssh 1 conns -> 192.168.1.233:2343
     - 13349/gmpc 1 conns -> 55.198.52.5:80
     - 13349/gmpc 1 conns -> 890.18.221.162:80

The conscious reader will spot that the below example is from my
laptop, so the process names are obvious different from that of a
server, but the concept is _exactly_ the same, nevertheless.

2) Furthermore, if blockdiag is installed and the current user has
write access to /var/www/system-info, it will also generate a diagram
based on the information obtained in 1) giving an up to date
architectural overview of the current server!

  Current archicture overview:
     |
      cochise's diagram: http://cochise:5678/cochise.png

Generating the PNG takes some additional seconds so some folks will
probably want to turn this off, in which case the following setting in
/etc/default/system-info is supported:

  do_not_generate_diagram=1