Made it possible to run ece-install with no "technet download list".
The list of artifacts needed for installation of a machine can be
specified by the EAR file.  The EAR file should contain an "effective
pom" with normal <properties> in a specific form.  ece-install looks
through the repositories specified by fai_maven_repositories for the
artifacts and downloads them in addition to wf_download_list and
technet_download_list (which may be empty).

Assemblytool is no longer required to run ece-install with an EAR.