VOSA-228: publication webapps now get access logs again
This is a regression of 702368fc8b8ed80e368fe172c44cdb68e24d28bc when
a separate HTTP connector for the publication webapps was set
up. Since the access log valve has the same prefix as the one defined
for the defualt HTTP connector, the second valve wasn't working. As a
result, access log entries for the publication webapps were no longer
available on the system.