ece deploy --file now caches EAR file like it does with --uri
When running 'ece deploy --file <file>', a symlink to the file was
created. However, this defies the point of the cache directory
somewhat, since the cache dir is for providing data that the ece
command can re-use on later runs.

When the file in /var/cache/escenic is only a symlink, there's no
guarantee that this file can be re-used. Moreover, since this file
often resides in /tmp, it will be deleted once the machine is
restarted.

This change copies the EAR file to the cache directory, making its
behaviour consistent with what happens when ece deploy is used with an
EAR residing on a remote machine.