If no EAR is passed, repackage is now idempotent
If the user didn't pass any EAR to repackage, we want, if possible, to
pick the previous EAR and re-package that. This is because we want to
be able to make apt-get install/upgrade/remove idempotent, not to
repackage a repackaged EAR.

Goal: apt-get install <package> && apt-get remove <package> should
leave you in the same step as before you started.