ece -i search1 -t search deploy now works with EARs without engine-config
Search instances should not have the engine-config JAR as this will
bootstrap the Nursery framework, which the search doesn't need. This will
just generate loads of errors in the logs.

deploy would therefore remove this JAR file. However, if the EAR doesn't
contain this config jar, the deployment should still work. This fix
fulfils this scenario  having an EAR file tailored for the search
instance.