With Varnish 4, there's no reason for date hack
In Varnish 3, date was incorrectly always set to "now" making it hard
to do the right thing with the Age header, setting it to 0 was our
only solution, see
https://www.varnish-cache.org/lists/pipermail/varnish-commit/2014-April/011368.html

However, this has been fixed in Varnish 4, so that the date header
from origin is kept. This means we can remove the hard coded Age: 0
response header.